backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "小红书热点 Agent 后端"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "aiosmtplib>=3.0.0",
    "apscheduler>=3.10.0,<4.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "markdown>=3.5",
    "openai>=2.28.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
